requirements.txt
numpy==1.22.0
pandas==1.4.2
matplotlib==3.5.1
seaborn==0.11.2
requests==2.27.1
jupyter==1.0.0
scikit-learn==1.0.2
spacy==3.2.4
nltk==3.7
elasticsearch==8.1.2
tensorflow==2.8.0
keras==2.8.0
python-dotenv==0.20.0
loguru==0.6.0
minio==7.1.5
vk-api==11.9.7